Refactor project to have a clear structure
This is using folder-by-feature structure.
